Implemented time range for stacky.py search
Stacktach support when_min and when_max as filter arguments for GET
/stacky/search/ . This patch implements support to pass when_min
and when_max in stacky.py search option.
The when_min and when_max datetime are accepted in friendly format
as seen with results of stacky.py search and other options.

Added examples on how to use time range for 'search' feature of stacky
on README.md